refactor(router): move lifecycle stuff into its own folder
