+++ modules/angular2/src/router/lifecycle/route_lifecycle_reflector.dart
@@ -1,6 +1,6 @@
 library angular.router.route_lifecycle_reflector;
 
-import 'package:angular2/src/router/lifecycle_annotations_impl.dart';
+import 'package:angular2/src/router/lifecycle/lifecycle_annotations_impl.dart';
 import 'package:angular2/src/router/interfaces.dart';
 import 'package:angular2/src/core/reflection/reflection.dart';
