+++ modules/angular2/src/router/lifecycle/route_lifecycle_reflector.ts
