+++ modules/angular2/test/router/integration/lifecycle_hook_spec.ts
@@ -40,7 +40,7 @@ import {
   CanDeactivate,
   CanReuse
 } from 'angular2/src/router/interfaces';
-import {CanActivate} from 'angular2/src/router/lifecycle_annotations';
+import {CanActivate} from 'angular2/src/router/lifecycle/lifecycle_annotations';
 import {ComponentInstruction} from 'angular2/src/router/instruction';
